HomeWork Lesson 17. Work auth with module auth.js. Kulik Vyacheslav
