+++ Vyacheslav_Kulik/lesson 17/auth.js
@@ -3,37 +3,42 @@ const LocalStrategy = require('passport-local').Strategy
 
 let Users = require('./models/users')
 
-passport.use(new LocalStrategy(
-    {
-        usernameField: 'email'
-    },
-    function(username, password, done) {
 
-        Users.findOne({ 'email': username }, function (err, user) {
-            if (err) { return done(err); }
-            if (!user) {
-                console.log('this log func')
-                return done(null, false, { message: 'Incorrect username.' });
+passport.use('local', new LocalStrategy(
+    {
+        usernameField: 'email_address' // поле формы авторизации, которое нужно использовать в качестве username
+    },
+    function (username, password, done) {
+
+        Users.findOne({email: username}, function (err, user) {
+            if (err) {
+                return done(err);
             }
-            if (!user.validatePassword(password)) {
-                return done(null, false, { message: 'Incorrect password.' });
-            }
-            return done(null, user)
+            // if (!user) { \\ эта ветка не используется, так как я реализовал ее через валидацию (express-validator)
+            //     return done(null, false);
+            // }
+            // if (!user.validatePassword(password)) { \\ эта ветка не используется, так как я реализовал ее через валидацию (express-validator)
+            //     console.log('validatePassword!!!!!')
+            //     return done(null, false);
+            // }
+            const plainUser = JSON.parse(JSON.stringify(user))
+            delete plainUser.password
+            return done(null, plainUser)
         })
     }
 ))
 
-passport.serializeUser(function(user, done) {
-    console.log(user)
-    done(null, user.id)
+passport.serializeUser(function (user, done) { //после процедуры авторизации поместить в куки данные о пользователе (его id)
+    console.log(user, 'serializeUser')
+    done(null, user._id) // - в куках выглядит вот так "passport":{"user":"5f50a39603361338b0cd067c"}
 })
 
-passport.deserializeUser(function(id, done) {
-    Users.findById(id, function(err, user) {
-        console.log(user)
+passport.deserializeUser(function (id, done) { //берется из записи сесии в BD  (passport":{"user":"5f50a39603361338b0cd067c"}) id = passport.user
+    Users.findById(id, function (err, user) {
         const plainUser = JSON.parse(JSON.stringify(user))
         delete plainUser.password
-        done(err, plainUser);
+        //console.log(plainUser, 'deserializeUser plainUser')
+        done(err, plainUser); //  при logout  запись из сесии в БД удаляется - "passport":{}
     })
 })
 
@@ -41,19 +46,26 @@ passport.deserializeUser(function(id, done) {
 module.exports = {
     initialize: passport.initialize(),
     session: passport.session(),
-
     //Обработчик для auth
     authenticate: passport.authenticate('local', {
-        successRedirect: '/todo',
-        failureRedirect: '/auth?error=1',
+        session: true,
+        successRedirect: '/todo'
     }),
-
-    //Проверяем возможность доступа
-    isAuthenticated: (req, res, next) => {
-        if(req.user){
-            next()
+    //защита от авторизованных пользователей
+    checkAuth: function (req, res, next) {
+        if (req.isAuthenticated()) {
+            res.redirect('/todo')
         } else {
+            next()
+        }
+    },
+    //защита от не авторизованных пользователей
+    checkNotAuth: function (req, res, next) {
+        if (!req.isAuthenticated()) { //защита от не авторизированых пользователей
+            res.cookie('notFirstUseSite', 'true')
             res.redirect('/auth')
+        } else {
+            next()
         }
     },
